+++ fenecon2mqtt/rootfs/usr/bin/fenecon2mqtt/config.py
@@ -1,13 +1,17 @@
 #!/usr/bin/python3
 import json
+import os
+import logging
 
 data = None
 
 try:
     with open('/data/options.json') as file:
         data = json.load(file)
-except e:
-    with open('config.json') as file:
+except Exception:
+    script_dir = os.path.dirname(__file__)
+    rel_path = "./config.json"
+    with open(os.path.join(script_dir, rel_path)) as file:
         data = json.load(file)
 
 fenecon = data['fenecon']